hair_project/requirements.txt
streamlit>=1.28.0
django>=5.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.21.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
scipy>=1.11.0
xformers>=0.0.20